snapshot require optimizations/fixes
