+++ snapshot/lib/snapshot/detect_values.rb
@@ -1,3 +1,6 @@
+require 'fastlane_core/project'
+require 'fastlane_core/device_manager'
+
 module Snapshot
   class DetectValues
     # This is needed as these are more complex default values
