+++ snapshot/lib/snapshot/simulator_launchers/launcher_configuration.rb
@@ -1,4 +1,3 @@
-
 module Snapshot
   class SimulatorLauncherConfiguration
     # both
