+++ snapshot/lib/snapshot/test_command_generator_base.rb
@@ -1,3 +1,6 @@
+require 'fastlane_core/device_manager'
+require_relative 'module'
+
 module Snapshot
   class TestCommandGeneratorBase
     class << self
