+++ snapshot/lib/snapshot/test_command_generator_xcode_8.rb
@@ -1,4 +1,6 @@
-require 'snapshot/test_command_generator_base'
+require_relative 'test_command_generator_base'
+require_relative 'module'
+require_relative 'latest_os_version'
 
 module Snapshot
   # Responsible for building the fully working xcodebuild command
